+++ presentation/announcement/src/main/java/com/saegil/announcement/announcement/AnnouncementScreen.kt
@@ -1,4 +1,4 @@
-package com.seagil.mypage.mypage
+package com.saegil.announcement.announcement
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -10,22 +10,19 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 
 @Composable
-fun MypageScreen(
-//    state: MypageState,
-//    actions: MypageActions
+fun AnnouncementScreen(
 ) {
     // TODO UI Rendering
     Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-        Text(text = "마이페이지 화면", fontSize = 24.sp)
+        Text(text = "공지 화면", fontSize = 24.sp)
     }
 }
 
+
 @Composable
-@Preview(name = "Mypage")
-private fun MypageScreenPreview() {
-    MypageScreen(
-//        state = MypageState(),
-//        actions = MypageActions()
+@Preview(name = "Announcement")
+private fun AnnouncementScreenPreview() {
+    AnnouncementScreen(
     )
 }
